Send med gyldige hjemler
